app/src/main/java/com/indie/apps/pannypal/domain/usecase/addPaymentUsecase.kt
package com.indie.apps.pannypal.domain.usecase

import android.net.http.HttpException
import androidx.compose.ui.unit.Constraints
import com.indie.apps.pannypal.data.dao.MerchantDataDao
import com.indie.apps.pannypal.data.entity.Merchant
import com.indie.apps.pannypal.data.entity.MerchantData
import com.indie.apps.pannypal.data.entity.Payment
import com.indie.apps.pannypal.di.IoDispatcher
import com.indie.apps.pannypal.data.module.MerchantDataWithName
import com.indie.apps.pannypal.repository.MerchantDataRepository
import com.indie.apps.pannypal.repository.MerchantRepository
import com.indie.apps.pannypal.repository.PaymentRepository
import com.indie.apps.pannypal.repository.UserRepository
import com.indie.apps.pannypal.util.Constant
import com.indie.apps.pannypal.util.Resource
import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flow
import kotlinx.coroutines.flow.flowOn
import java.io.IOException
import javax.inject.Inject

class addPaymentUsecase @Inject constructor(
    private val paymentRepository: PaymentRepository,
    private val payment: Payment,
    @IoDispatcher private val dispatcher: CoroutineDispatcher) {

    suspend operator fun invoke() : Flow<Resource<Long>>{
        return flow{

            try {
                emit(Resource.Loading<Long>())
                val id = paymentRepository.insert(payment)

                if(id >0)
                {
                    emit(Resource.Success<Long>(id))
                }else{
                    emit(Resource.Error<Long>("Payment not add"))
                }


            } catch(e: Throwable) {
                when(e) {
                    is IOException -> emit(Resource.Error<Long>("Network Failure"))
                    else -> emit(Resource.Error<Long>("Conversion Error"))
                }
            }
        }.flowOn(dispatcher)
    }

}